requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
tensorflow>=2.16.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
